pyproject.toml
[tool.poetry]
name = "web-comments"
version = "0.1.0"
description = ""
authors = ["BRANYA43 <tbd2504@gmai.com>"]
license = "MIT"
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
channels = {extras = ["daphne"], version = "^4.2.0"}
django-bootstrap5 = "^24.3"
psycopg = "^3.2.3"
django-split-settings = "^1.3.2"
django-cleanup = "^9.0.0"
pydantic-settings = "^2.6.1"
pillow = "^11.0.0"
pydantic = {extras = ["email"], version = "^2.10.3"}
drf-spectacular = "^0.28.0"
drf-standardized-errors = {extras = ["openapi"], version = "^0.14.1"}
django-filter = "^24.3"
djangochannelsrestframework = "^1.2.0"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.9.0"
pytest-selenium = "^4.1.0"
pre-commit = "^4.0.1"
pytest = "^8.3.4"
selenium = "^4.27.1"
pytest-lazy-fixtures = "^1.1.1"
faker = "^33.1.0"
pytest-factoryboy = "^2.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120
force-exclude = true
extend-exclude = ['migrations']
lint.extend-select = ['T20']

[tool.ruff.format]
quote-style = "single"


[tool.mypy]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ['*.migrations.*']
ignore_errors = true


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'core.settings'
pythonpath = 'src'
django_debug_mode = true
addopts = '--reuse-db'
